requirements.txt
python-dotenv==0.20.0
more-itertools==8.13.0
python-telegram-bot==13.2
